fix hien thi ten, load anh trong dang bai

+++ API/src/main/java/vn/agriconnect/API/dto/request/post/PostCreateRequest.java
@@ -12,24 +12,24 @@ import java.util.List;
 public class PostCreateRequest {
     @NotBlank(message = "Title is required")
     private String title;
-    
+
     private String description;
-    
+
     @NotNull(message = "Price is required")
     @Positive(message = "Price must be positive")
     private Double price;
-    
+
     @NotBlank(message = "Unit is required")
     private String unit;
-    
+
     @NotNull(message = "Quantity is required")
     @Positive(message = "Quantity must be positive")
-    private Integer quantity;
-    
+    private Double quantity;
+
     private List<String> images;
-    
+
     @NotBlank(message = "Category is required")
     private String categoryId;
-    
+
     private Location location;
 }
